# server/src/openapi/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Architecture Artifacts Core API",
    "description": "Core API endpoints for the Architecture Artifacts application - a comprehensive enterprise architecture documentation platform with multi-space file management, Git integration, authentication, templates, microservices, and search functionality.",
    "version": "2.0.0",
    "contact": {
      "name": "Design Artifacts Team"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Core API server"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Session-based authentication using cookies"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "FileTreeItem": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": ["file", "directory"]
          },
          "path": {
            "type": "string"
          },
          "fileType": {
            "type": "string",
            "enum": ["markdown", "pdf", "image", "text", "unknown"]
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FileTreeItem"
            }
          }
        }
      },
...
      "SearchResult": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "fileName": {
            "type": "string"
          },
          "filePath": {
            "type": "string"
          },
          "preview": {
            "type": "string"
          },
          "matchIndex": {
            "type": "integer"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/files": {
      "get": {
        "summary": "Get file tree",
        "tags": ["Files"],
        "responses": {
          "200": {
            "description": "File tree structure",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FileTreeItem"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create new markdown file",
        "tags": ["Files"],
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["filePath"],
                "properties": {
                  "filePath": {
                    "type": "string",
                    "description": "Path for the new file (must end with .md)"
                  },
                  "content": {
                    "type": "string",
                    "default": ""
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "path": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "File already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/search/files": {
      "get": {
        "summary": "Search files by name",
        "tags": ["Search"],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Search query"
          },
          {
            "name": "space",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filter search to specific space"
          }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchResult"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/search/content": {
      "get": {
        "summary": "Search content within files",
        "tags": ["Search"],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Search query"
          },
          {
            "name": "space",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filter search to specific space"
          }
        ],
        "responses": {
          "200": {
            "description": "Content search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchResult"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}